requirements.txt
# azure
azure-core
azure-ai-documentintelligence

# paddle
paddleocr
paddlepaddle

# classifier
timm

# runner
pdf2image
python-dotenv